requirements.txt
Django==2.0.3
djangorestframework==3.7.7
celery==4.1.0
django-celery-beat==1.1.1
xmltodict==0.10.2
uwsgi==2.0.17
redis==2.10.5
hiredis==0.2.0
mysqlclient==1.3.10
pymongo==3.4.0
requests==2.13.0
django-debug-toolbar==1.7
Pillow==4.0.0
pycrypto==2.6.1
django-redis-cache==1.7.1
django-prettyjson
jsonfield==2.0.2
numpy==1.12.0
aiohttp==3.1.2
pyjwt==1.6.1
ipython
user-agents==1.1.0
django-user-agents==0.3.0
django-aliyun-oss
django-admin-rangefilter
pyDes==2.0.1
jieba==0.39
scipy==1.1.0
baidu-aip